Revert "Add Polymorphism examples (#5)"
This reverts commit f43ec2281d3c8536a51feb0d7a7bd44c60be4712.
