--- src/com/OOP/Inheritance/Animal/animalTest.java
@@ -1,14 +1,0 @@
-package OOP.Inheritance.Animal;
-
-public class animalTest {
-    public static void main(String[] args) {
-        Cat cat = new Cat(false, "milk", 4);
-
-        System.out.println("Cat is vegetarian?" + cat.isVegetarian());
-        System.out.println("Cat eats " + cat.getEats());
-        System.out.println("Cat has " + cat.getNoOfLegs() + " legs.");
-        System.out.println("Cat color is " + cat.getColor());
-
-    }
-
-}
