--- src/com/OOP/Polymophism/Shapes/Square.java
@@ -1,15 +1,0 @@
-package OOP.Polymophism.Shapes;
-
-class Square extends Rectangle{
-    private double side;
-
-    Square(double side){
-        super(side);
-    }
-
-    @Override
-    public void calculateArea(){
-        System.out.println("Square Area = " + side * side);
-    }
-
-}
