+++ src/com/Tutorial/Tutorial_3/Tute3ex3/DemoApp.java
@@ -1,0 +1,9 @@
+package Tutorial.Tutorial_3.Tute3ex3;
+
+class DemoApp {
+    public static void main(String[] args) {
+        Calc.addition(3, 2);
+        Calc.subtraction(5, 3);
+    }
+}
+
