--- src/com/Tutorials/Tutorial_3/Tute3ex3/Calc.java
@@ -1,11 +1,0 @@
-package Tutorials.Tutorial_3.Tute3ex3;
-
-class Calc {
-    public static int addition(int n1, int n2) {
-        return n1 + n2;
-    }
-
-    public static int subtraction(int n1, int n2) {
-        return n1 - n2;
-    }
-}
