--- src/com/Tutorials/Tutorial_3/Tute3ex3/DemoApp.java
@@ -1,9 +1,0 @@
-package Tutorials.Tutorial_3.Tute3ex3;
-
-class DemoApp {
-    public static void main(String[] args) {
-        Calc.addition(3, 2);
-        Calc.subtraction(5, 3);
-    }
-}
-
